refactor(profile): refactor user data screen into separate component
